Implement 'arguments' for strict mode
If the identifier 'arguments' is used and it hasn't been declared,
a mapping to a 'js::Arguments' object is created. It holds a copy
of the actual arguments. This is only the strict mode interpretation.
Non-strict mode support requires modification of 'arguments' to be
propagated back to the actual argument. While we have an idea how
to do that, it is not important enough for us to bother with it
at this time.