runtime: BUGFIX: StackFrame marking itself before constructed
When allocating the environment in the stack frame, the frame
passed itself as "caller" to the allocator. If a GC occurred,
the allocator would try to mark the frame, which wasn't fully
constructed yet. The code did make an attempt to have the frame
in consistent state by zeroing the 'env' pointer, but the
locals hadn't been initialized.

All of this is unnecessary because the allocator needs the calling
frame, not the currently constructed one.