runtime: improve property name comparison
We are using std::map-s, so we have provided less() operators for the
couple of cases where we need to compare strings.

Improve the implementations of less() by first comparing the string
pointers for equality. If the strings really are the same, the
comparison would take the longest as it has to reach the end of both
strings.