Groundwork for "true" bind
Every Function object has a pointer to a function and a pointer to a
constructor, which can optionally be different from the function. The
latter is used only by "new". This will allow our "native" bind to know
how it is being called.

Also, this simplified and cleaned-up some system constructors (Number,
Boolean, etc).