runtime: free unused interned strings
We need to free unused interned strings, or our hash table is
unbounded (since we want all properties to be interned and there is no
fixed bounds on the possible properties).
Some strings can be marked as "permanent" so they are never freed.
This is necessary for the literals - it would be to painful to have to
mark all literals.