runtime: add NativeObject for native interfaces
NativeObject contains a configurable number of "internal" (or native)
properties of type uintptr_t and has an optional finalizer.

For example a native file handle can be stored in an internal
property and it can be closed in the finalizer.

Note that this is not a full blown finalizer in the conventional sense.
It is a light-weight native function that cannot resurrect the object
and gets executed immediately in the garbage collector.