runtime: remove all instances of dynamic_cast
Instead use getInternalClass(). It is a virtual function for now, but
when we transition to our own classes, it will be just a field in the
class, so accessing it will be even faster.

This lead to a 20% spead up in gc_collect(), which used dynamic_cast
to check for interned strings...