compiler, runtime: builtin modules are descibed in the runtime
Instead of manually importing the builting modules (process, console)
in the compiler, we define them in the runtime. The only trick is
that the actual "require" call has to be delayed until after the
runtime has initialized, so we define a new special runtime function
"runtimeInit" which we call manually after the runtime has initialized.
That's where the "require" calls are.