runtime: implement "init tags"
Init tags allow us to verify that a NativeObject was created and
initialized by us, as opposed to created by someone else who has access
to our prototype. It is a way to achieve reliable static typing.

An "init tag" is created by calling "$jsc.newInitTag()" (it is simply
an object, but could change in the future) and is stored in a private
location. Later, when a new instance is initialized, the tag is applied
to the instance with "$jsc.setInitTag(obj, tag)". Only we can do that
because only we have access to the tag.

To verify that an object is of the correct type, we use
"$jsc.checkInitTag(obj, tag)". Since only we have access to the tag
value and only we could have applied it to an object, this is a robust
way that the object is what we think it is.